--- a/Cargo.toml
+++ b/Cargo.toml
@@ -20,6 +20,7 @@ path = "src/main.rs"
 
 [dependencies]
 ansi_term = "0.12.0"
+anyhow = "1"
 app_dirs = { version = "2", package = "app_dirs2" }
 atty = "0.2"
 clap = { version = "3", features = ["std", "derive", "suggestions", "color"], default-features = false }