--- a/src/error.rs
+++ b/src/error.rs
@@ -2,11 +2,11 @@ use std::fmt;
 
 use reqwest::Error as ReqwestError;
 
+
 #[derive(Debug)]
 #[allow(clippy::enum_variant_names)]
 pub enum TealdeerError {
     CacheError(String),
-    ConfigError(String),
     UpdateError(String),
     WriteError(String),
 }
@@ -15,7 +15,6 @@ impl TealdeerError {
     pub fn message(&self) -> &str {
         match self {
             Self::CacheError(msg)
-            | Self::ConfigError(msg)
             | Self::UpdateError(msg)
             | Self::WriteError(msg) => msg,
         }
@@ -32,7 +31,6 @@ impl fmt::Display for TealdeerError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
             Self::CacheError(e) => write!(f, "CacheError: {}", e),
-            Self::ConfigError(e) => write!(f, "ConfigError: {}", e),
             Self::UpdateError(e) => write!(f, "UpdateError: {}", e),
             Self::WriteError(e) => write!(f, "WriteError: {}", e),
         }