--- a/src/error.rs
+++ b/src/error.rs
@@ -2,23 +2,10 @@ use std::fmt;
 
 use reqwest::Error as ReqwestError;
 
-
 #[derive(Debug)]
 #[allow(clippy::enum_variant_names)]
 pub enum TealdeerError {
-    CacheError(String),
     UpdateError(String),
-    WriteError(String),
-}
-
-impl TealdeerError {
-    pub fn message(&self) -> &str {
-        match self {
-            Self::CacheError(msg)
-            | Self::UpdateError(msg)
-            | Self::WriteError(msg) => msg,
-        }
-    }
 }
 
 impl From<ReqwestError> for TealdeerError {
@@ -30,9 +17,7 @@ impl From<ReqwestError> for TealdeerError {
 impl fmt::Display for TealdeerError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
-            Self::CacheError(e) => write!(f, "CacheError: {}", e),
             Self::UpdateError(e) => write!(f, "UpdateError: {}", e),
-            Self::WriteError(e) => write!(f, "WriteError: {}", e),
         }
     }
 }