--- a/src/utils.rs
+++ b/src/utils.rs
@@ -12,6 +12,17 @@ pub fn print_error(enable_styles: bool, message: &str) {
     print_msg(enable_styles, message, "Error: ", Color::Red);
 }
 
+/// Print an anyhow error to stderr. If `enable_styles` is true, then a red
+/// message will be printed.
+pub fn print_anyhow_error(enable_styles: bool, error: &anyhow::Error) {
+    print_msg(
+        enable_styles,
+        &format!("{:?}", error),
+        "Error: ",
+        Color::Red,
+    );
+}
+
 fn print_msg(enable_styles: bool, message: &str, prefix: &'static str, color: Color) {
     if enable_styles {
         let style = Style::new().fg(color);