Use [anyhow](https://docs.rs/anyhow/) instead of custom error types and strings for error handling. This should give us "caused by" style traces when an error occurs.

Before:

```
$ tldr --update
Error: Could not update cache: Could not remove cache directory (/home/danilo/.cache/tealdeer/tldr-pages): Permission denied (os error 13)
```

And after:

```
$ cargo run -- --update
Error: Could not update cache

Caused by:
    0: Could not clear the cache directory
    1: Could not remove the cache directory at /home/danilo/.cache/tealdeer/tldr-pages
    2: Permission denied (os error 13)
```

PR summary:

>  116 additions and 195 deletions

We're doing something right 😉