@niklasmohrin 2022-01-08
It might also be worth it to inspect all uses of `unwrap` etc. and check whether it is now more reasonable to propagate as an anyhow Erorr instead