@dbrgn 2022-01-08
> It might also be worth it to inspect all uses of `unwrap` etc. and check whether it is now more reasonable to propagate as an anyhow Erorr instead (maybe in another issue / PR though)

I didn't really find problematic unwraps, except for that unzip issue that popped up in that other issue. However, let's fix that separately.