@dbrgn 2022-01-08
> I am building the benchmark docker right now!

Let me know if it makes a difference! The only place that _could_ make a difference in my eyes is the replacement of the `unwrap` in the inner printing loop with a `.context(...)?`. However, we're not exactly dealing with millions of loop iterations, so I don't think it'll make a significant difference.